// packages/apollo-wind-css/src/generated/css-vars.ts
export const cssVars = {
  color: {
    palette: {
// ...
      'blue-secondary': {
        '100': '--color-blue-secondary-100',
        '200': '--color-blue-secondary-200',
        '300': '--color-blue-secondary-300',
        '400': '--color-blue-secondary-400',
        '500': '--color-blue-secondary-500',
        '600': '--color-blue-secondary-600',
        '700': '--color-blue-secondary-700',
        '800': '--color-blue-secondary-800',
      },
// ...
      'light-blue': {
        '100': '--color-light-blue-100',
        '200': '--color-light-blue-200',
        '300': '--color-light-blue-300',
        '400': '--color-light-blue-400',
        '500': '--color-light-blue-500',
        '600': '--color-light-blue-600',
        '700': '--color-light-blue-700',
        '800': '--color-light-blue-800',
      },
// ...
    },
// ...
  },
// ...
} as const;